1. 调通音频播放，SDL_OpenAudio(&wantedSpec, NULL); 第二个参数设置为NULL,  官方解释   If \c obtained is NULL, the audio  *  data passed to the callback function will be guaranteed to be in the  *  requested format
2. uint8_t 和 uint8_t * 的值在Qt环境调试器上显示的是 0，所以尽量改为int或者uint
3. 主函数中增加音频输出设备列表打印

+++ player/MyPlayer.h
@@ -32,7 +32,7 @@ public:
     void InitAVFrameQueue();
     void StreamComponentOpen(int streamIndex);   // 根据流索引得到流AVStream和打开相关组件
     void OpenAudioDevice();
-    static void SdlAudioCallBack(void *userdata, uint8_t *stream, int needLen);
+    static void SdlAudioCallBack(void *userdata, Uint8 *stream, int needLen);
     int DecodeAudioFrame();
 
     AvMsgQueue m_queue;
@@ -51,8 +51,8 @@ public:
     AVPacket *m_audioPacket;
     AVFrame *m_audioFrame;
     uint8_t *m_pAudioBuffer = NULL;
-    uint8_t m_audioBufferSize = 0;      // 音频缓冲区大小字节数, 此处其实不用太纠结索引从0还是1开始，因为一个字节不会影响音频播放
-    uint8_t m_audioBufferUsedSize = 0;  // 缓冲区已经使用的音频字节数
+    unsigned int m_audioBufferSize = 0;      // 音频缓冲区大小字节数, 此处其实不用太纠结索引从0还是1开始，因为一个字节不会影响音频播放
+    unsigned int m_audioBufferUsedSize = 0;  // 缓冲区已经使用的音频字节数
 
     struct SwrContext* m_audioSwrCtx = NULL;   // 音频重采样上下文
